dashboard migration: refactor
