+++ public/app/core/constants.ts
@@ -4,4 +4,5 @@ export const GRID_CELL_VMARGIN = 10;
 export const GRID_COLUMN_COUNT = 24;
 export const REPEAT_DIR_VERTICAL = 'v';
 
-
+export const DEFAULT_ROW_HEIGHT = 250;
+export const MIN_PANEL_HEIGHT = GRID_CELL_HEIGHT * 3;
